cleaning up folder structure
