+++ project-folder/src/pages/OldHome.jsx
@@ -1,0 +1,18 @@
+import React from 'react'
+import UserNavbar from '../components/Navbar.jsx'
+import Brain from '../brain-main-old/Brain.jsx'
+import Home from './Home.jsx'
+import Footer from "../components/Footer.jsx";
+import AboutUs from "./AboutUs.jsx";
+
+export default function OldHome() {
+
+    return (
+        <>
+            <Home />
+            {/* <Brain /> */}
+            {/*<AboutUs/>*/}
+            <Footer />
+        </>
+    )
+}
